+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 
+## [1.9.0] - 2022-08-16
+
 ### Added
 
 - Added configuration options to enable certain Language Server features. By default, they are all enabled: